restructure in anticipation of further subcommands
