+++ Cargo.toml
@@ -10,23 +10,22 @@ homepage = "https://casperlabs.io"
 repository = "https://github.com/casper-network/casper-db-utils"
 license = "Apache-2.0"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 anyhow = "1"
-bincode = "1.3.3"
-casper-execution-engine = "2.0.0"
-casper-hashing = "1.4.3"
-casper-node = "1.4.5"
-casper-types = "1.5.0"
-clap = "3.1.17"
+bincode = "1"
+casper-execution-engine = "2"
+casper-hashing = "1"
+casper-node = "1"
+casper-types = "1"
+clap = "3"
 lmdb = "0.8.0"
 lmdb-sys = "0.8.0"
 log = "0.4.17"
-serde = {version = "1", features = ["derive"] }
+serde = { version = "1", features = ["derive"] }
 simplelog = "0.12.0"
-thiserror = "1.0.31"
+thiserror = "1"
 
 [dev-dependencies]
+once_cell = "1"
 rand = "0.8.5"
-tempfile = "3.3.0"
+tempfile = "3"
