+++ src/subcommands.rs
@@ -1,0 +1,2 @@
+pub mod check;
+pub mod trie_compact;
