+++ src/subcommands/check/db/block_body_merkle_db.rs
@@ -3,25 +3,26 @@ use std::{
     result::Result,
 };
 
+use casper_hashing::Digest;
 use casper_types::bytesrepr::FromBytes;
 
-use crate::db::{Database, DeserializationError};
+use super::{Database, DeserializationError};
 
-pub struct StateStoreDatabase;
+pub struct BlockBodyMerkleDatabase;
 
-impl Display for StateStoreDatabase {
+impl Display for BlockBodyMerkleDatabase {
     fn fmt(&self, f: &mut Formatter<'_>) -> FormatterResult {
-        write!(f, "state_store")
+        write!(f, "block_body_merkle")
     }
 }
 
-impl Database for StateStoreDatabase {
+impl Database for BlockBodyMerkleDatabase {
     fn db_name() -> &'static str {
-        "state_store"
+        "block_body_merkle"
     }
 
     fn parse_element(bytes: &[u8]) -> Result<(), DeserializationError> {
-        let _: u64 = FromBytes::from_bytes(bytes)?.0;
+        let _: (Digest, Digest) = FromBytes::from_bytes(bytes)?.0;
         Ok(())
     }
 }
