+++ src/subcommands/check/db/deploy_hashes_db.rs
@@ -5,7 +5,7 @@ use std::{
 
 use casper_types::{bytesrepr::FromBytes, DeployHash};
 
-use crate::db::{Database, DeserializationError};
+use super::{Database, DeserializationError};
 
 pub struct DeployHashesDatabase;
