+++ src/subcommands/check/db/finalized_approvals_db.rs
@@ -5,7 +5,7 @@ use std::{
 
 use casper_node::types::FinalizedApprovals;
 
-use crate::db::{Database, DeserializationError};
+use super::{Database, DeserializationError};
 
 pub struct FinalizedApprovalsDatabase;
