+++ src/subcommands/check/db/state_store_db.rs
@@ -3,26 +3,25 @@ use std::{
     result::Result,
 };
 
-use casper_hashing::Digest;
 use casper_types::bytesrepr::FromBytes;
 
-use crate::db::{Database, DeserializationError};
+use super::{Database, DeserializationError};
 
-pub struct BlockBodyMerkleDatabase;
+pub struct StateStoreDatabase;
 
-impl Display for BlockBodyMerkleDatabase {
+impl Display for StateStoreDatabase {
     fn fmt(&self, f: &mut Formatter<'_>) -> FormatterResult {
-        write!(f, "block_body_merkle")
+        write!(f, "state_store")
     }
 }
 
-impl Database for BlockBodyMerkleDatabase {
+impl Database for StateStoreDatabase {
     fn db_name() -> &'static str {
-        "block_body_merkle"
+        "state_store"
     }
 
     fn parse_element(bytes: &[u8]) -> Result<(), DeserializationError> {
-        let _: (Digest, Digest) = FromBytes::from_bytes(bytes)?.0;
+        let _: u64 = FromBytes::from_bytes(bytes)?.0;
         Ok(())
     }
 }
