+++ src/subcommands/trie_compact/tests.rs
@@ -1,5 +1,9 @@
 use std::fs::{self, File};
 
+use lmdb::DatabaseFlags;
+use once_cell::sync::Lazy;
+use tempfile::{tempdir, TempDir};
+
 use casper_execution_engine::storage::{
     store::StoreExt,
     transaction_source::{lmdb::LmdbEnvironment, Transaction, TransactionSource},
@@ -9,16 +13,12 @@ use casper_execution_engine::storage::{
 use casper_hashing::Digest;
 use casper_node::storage::Storage;
 use casper_types::bytesrepr::{Bytes, ToBytes};
-use lmdb::DatabaseFlags;
-use tempfile::{tempdir, TempDir};
 
-use crate::trie_compact::compact::DEFAULT_MAX_DB_SIZE;
+static DEFAULT_MAX_DB_SIZE: Lazy<usize> = Lazy::new(|| super::DEFAULT_MAX_DB_SIZE.parse().unwrap());
 
 use super::{
-    compact::TRIE_STORE_FILE_NAME,
-    trie_compact,
+    compact::{self, DestinationOptions, Error, TRIE_STORE_FILE_NAME},
     utils::{create_execution_engine, create_storage, load_execution_engine},
-    DestinationOptions, Error,
 };
 
 #[derive(Clone, Debug, PartialEq)]
@@ -92,7 +92,7 @@ fn create_data() -> Vec<TestData<Bytes, Bytes>> {
 
 fn create_test_trie_store() -> (TempDir, Vec<TestData<Bytes, Bytes>>) {
     let tmp_dir = tempdir().unwrap();
-    let env = LmdbEnvironment::new(&tmp_dir.path(), DEFAULT_MAX_DB_SIZE, 512, true).unwrap();
+    let env = LmdbEnvironment::new(&tmp_dir.path(), *DEFAULT_MAX_DB_SIZE, 512, true).unwrap();
     let store = LmdbTrieStore::new(&env, None, DatabaseFlags::empty()).unwrap();
     let data = create_data();
 
@@ -118,7 +118,7 @@ fn copy_state_root_roundtrip() {
     let src_tmp_dir = tempdir().unwrap();
     let dst_tmp_dir = tempdir().unwrap();
     let src_env =
-        LmdbEnvironment::new(&src_tmp_dir.path(), DEFAULT_MAX_DB_SIZE, 512, true).unwrap();
+        LmdbEnvironment::new(&src_tmp_dir.path(), *DEFAULT_MAX_DB_SIZE, 512, true).unwrap();
     let src_store = LmdbTrieStore::new(&src_env, None, DatabaseFlags::empty()).unwrap();
     // Construct mock data.
     let data = create_data();
@@ -133,14 +133,14 @@ fn copy_state_root_roundtrip() {
 
     let (source_state, _env) = load_execution_engine(
         src_tmp_dir.path(),
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
         Digest::default(),
         true,
     )
     .unwrap();
 
     let (destination_state, dst_env) =
-        create_execution_engine(dst_tmp_dir.path(), DEFAULT_MAX_DB_SIZE, true).unwrap();
+        create_execution_engine(dst_tmp_dir.path(), *DEFAULT_MAX_DB_SIZE, true).unwrap();
 
     // Copy from `node1`, the root of the created trie. All data should be copied.
     super::helpers::copy_state_root(data[3].0, &source_state, &destination_state).unwrap();
@@ -178,7 +178,7 @@ fn check_no_extra_tries() {
     let src_tmp_dir = tempdir().unwrap();
     let dst_tmp_dir = tempdir().unwrap();
     let src_env =
-        LmdbEnvironment::new(&src_tmp_dir.path(), DEFAULT_MAX_DB_SIZE, 512, true).unwrap();
+        LmdbEnvironment::new(&src_tmp_dir.path(), *DEFAULT_MAX_DB_SIZE, 512, true).unwrap();
     let src_store = LmdbTrieStore::new(&src_env, None, DatabaseFlags::empty()).unwrap();
     // Construct mock data.
     let data = create_data();
@@ -193,14 +193,14 @@ fn check_no_extra_tries() {
 
     let (source_state, _env) = load_execution_engine(
         src_tmp_dir.path(),
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
         Digest::default(),
         true,
     )
     .unwrap();
 
     let (destination_state, dst_env) =
-        create_execution_engine(dst_tmp_dir.path(), DEFAULT_MAX_DB_SIZE, true).unwrap();
+        create_execution_engine(dst_tmp_dir.path(), *DEFAULT_MAX_DB_SIZE, true).unwrap();
 
     // Check with `node2`, which only has `leaf1` and `leaf2` as children in the constructed trie.
     super::helpers::copy_state_root(data[4].0, &source_state, &destination_state).unwrap();
@@ -248,12 +248,12 @@ fn check_no_extra_tries() {
 
 #[test]
 fn missing_source_trie() {
-    match trie_compact(
+    match compact::trie_compact(
         "".into(),
         "bogus_path".into(),
         "".into(),
         DestinationOptions::New,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::InvalidPath(..)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
@@ -265,12 +265,12 @@ fn missing_source_trie() {
 fn missing_storage() {
     let (src_dir, _) = create_test_trie_store();
     let dst_dir = tempdir().unwrap();
-    match trie_compact(
+    match compact::trie_compact(
         "bogus_path".into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::New,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::OpenStorage(_)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
@@ -283,36 +283,36 @@ fn valid_empty_dst_with_destination_options() {
     let (src_dir, _) = create_test_trie_store();
     let dst_dir = tempdir().unwrap();
     let (storage_dir, _store) = create_empty_test_storage();
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::New,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Ok(_) => {}
         Err(err) => panic!("Unexpected error: {}", err),
     }
     fs::remove_file(dst_dir.path().join(TRIE_STORE_FILE_NAME)).unwrap();
 
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::Append,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::InvalidDest(_)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
         Ok(_) => panic!("Unexpected successful trie compact"),
     }
 
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::Overwrite,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::InvalidDest(_)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
@@ -330,36 +330,36 @@ fn valid_existing_dst_with_destination_options() {
     }
 
     let (storage_dir, _store) = create_empty_test_storage();
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::New,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::InvalidDest(_)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
         Ok(_) => panic!("Unexpected successful trie compact"),
     }
 
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::Append,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Ok(_) => {}
         Err(err) => panic!("Unexpected error: {}", err),
     }
 
     assert!(dst_dir.path().join(TRIE_STORE_FILE_NAME).exists());
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.path().into(),
         DestinationOptions::Overwrite,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Ok(_) => {}
         Err(err) => panic!("Unexpected error: {}", err),
@@ -375,36 +375,36 @@ fn missing_dst_with_destination_options() {
     let root_dst_dir = tempdir().unwrap();
     let dst_dir = root_dst_dir.path().join("extra_dir");
     let (storage_dir, _store) = create_empty_test_storage();
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.clone(),
         DestinationOptions::New,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Ok(_) => {}
         Err(err) => panic!("Unexpected error: {}", err),
     }
     fs::remove_dir_all(dst_dir.as_path()).unwrap();
 
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir.clone(),
         DestinationOptions::Append,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::InvalidDest(_)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
         Ok(_) => panic!("Unexpected successful trie compact"),
     }
 
-    match trie_compact(
+    match compact::trie_compact(
         storage_dir.path().into(),
         src_dir.path().into(),
         dst_dir,
         DestinationOptions::Overwrite,
-        DEFAULT_MAX_DB_SIZE,
+        *DEFAULT_MAX_DB_SIZE,
     ) {
         Err(Error::InvalidDest(_)) => {}
         Err(err) => panic!("Unexpected error: {}", err),
